renamed package to greylock, moved tests folder to inside src, other minor changes
