+++ pyproject.toml
@@ -3,12 +3,12 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "metacommunity_diversity"
+name = "greylock"
 authors = [
     { name = "Elliot Hill", email = "elliot.douglas.hill@gmail.com" },
-    { name = "Jasper Braun" },
+    { name = "Jasper Braun", email = "jasperbraun90@gmail.com" },
 ]
-version = "0.0.1"
+version = "0.0.5"
 description = "A Python package for measuring the composition of complex datasets"
 readme = "README.md"
 requires-python = ">=3.8"
@@ -25,11 +25,12 @@ lint = ["black>=22.3.0"]
 ci = ["tox>=3.24.5", "tox-gh-actions>=2.9.1"]
 
 [project.urls]
-"Homepage" = "https://github.com/Elliot-D-Hill/diversity"
-"Bug Tracker" = "https://github.com/Elliot-D-Hill/diversity/issues"
+"Homepage" = "https://github.com/ArnaoutLab/greylock"
+"Bug Tracker" = "https://github.com/ArnaoutLab/greylock/issues"
 
 [tool.pytest.ini_options]
-addopts = "--cov --cov-report html --cov-report term-missing --cov-fail-under 100"
+addopts = ["--cov --cov-report html --cov-report term-missing --cov-fail-under 100", 
+"--import-mode=importlib"]
 
 [tool.coverage.run]
 source = ["src"]
