+++ src/greylock/metacommunity.py
@@ -11,12 +11,12 @@ from typing import Callable, Iterable, Optional, Union
 
 from pandas import DataFrame, Index, concat
 from numpy import atleast_1d, broadcast_to, divide, zeros, ndarray
-from metacommunity_diversity.exceptions import InvalidArgumentError
+from greylock.exceptions import InvalidArgumentError
 
-from metacommunity_diversity.abundance import make_abundance
-from metacommunity_diversity.similarity import make_similarity
-from metacommunity_diversity.components import make_components
-from metacommunity_diversity.utilities import power_mean
+from greylock.abundance import make_abundance
+from greylock.similarity import make_similarity
+from greylock.components import make_components
+from greylock.utilities import power_mean
 
 
 class Metacommunity:
