+++ src/greylock/tests/metacommunity_test.py
@@ -9,13 +9,13 @@ from numpy import (
 from pandas import DataFrame, concat
 from pandas.testing import assert_frame_equal
 from pytest import mark, raises
-from metacommunity_diversity.exceptions import InvalidArgumentError
+from greylock.exceptions import InvalidArgumentError
 
-from metacommunity_diversity.log import LOGGER
-from metacommunity_diversity.abundance import Abundance
-from metacommunity_diversity.similarity import Similarity
-from metacommunity_diversity import Metacommunity
-from tests.similarity_test import similarity_dataframe_3by3
+from greylock.log import LOGGER
+from greylock.abundance import Abundance
+from greylock.similarity import Similarity
+from greylock import Metacommunity
+from greylock.tests.similarity_test import similarity_dataframe_3by3
 
 MEASURES = (
     "alpha",
